Refactoer Sections and project Overviews
