+++ frontend/src/pages/ProjectDetails.jsx
@@ -9,6 +9,8 @@ import relativeTime from "dayjs/plugin/relativeTime";
 import PageTopic from "../components/PageTopic";
 import LoadingSpinner from "../components/LoadingSpinner";
 import PositioningMode from "../components/PositioningMode";
+import ProjectOverview from "../components/ProjectOverview";
+import SectionsOverview from "../components/SectionsOverview";
 
 // Extend dayjs with relativeTime
 dayjs.extend(relativeTime);
@@ -17,30 +19,14 @@ const ProjectDetails = () => {
   const { navigate, backendUrl, removeProject, fetchProject,project, setProject, updateProjectSections, removeProjectSection} = useContext(Context);
   const { projectId } = useParams();
   
-  
 
   const [exportFormat, setExportFormat] = useState("dwg");
   const [isExporting, setIsExporting] = useState(false);
   const [exportStatus, setExportStatus] = useState(null);
 
-  const [newSection, setNewSection] = useState("");
+  
 
-  const handleAddSection = () => {
-    const trimmed = newSection.trim();
-    if (trimmed === "") return;
-
-    if (project.Sections.includes(trimmed)) {
-      toast.error("Section already exists");
-    } else {
-      setProject({
-        ...project,
-        Sections: [...project.Sections, trimmed],
-      });
-      updateProjectSections(projectId, [...project.Sections, trimmed]);
-      setNewSection("");
-    }
-  };
-
+  
   const handleDelete = async () => {
     const ok = window.confirm(
       `Delete project “${project.Name}”?\nThis action cannot be undone.`
@@ -52,20 +38,7 @@ const ProjectDetails = () => {
     navigate(-1); // Go back
   };
 
-  const removeSection = async (projectId, section) => {
-    try {
-
-      await removeProjectSection(projectId, section);
-
-      if (response.data.success) {  
-        toast.success("Section removed successfully");
-      } else {
-        toast.error("Failed to remove section");
-      }
-    } catch (error) {
-      console.error("Error removing section:", error);
-    }
-  };
+  
 
   useEffect(() => {
     fetchProject(projectId);
@@ -298,66 +271,7 @@ const handleExport = async () => {
 
       <div className="flex flex-col gap-4">
          {/* Overview Section (Right) */}
-        <div className=" bg-white p-5 rounded-lg flex flex-col gap-3 h-max ">
-          <h2 className="text-base md:text-lg font-semibold pb-2">Overview</h2>
-
-          <div className="flex justify-between items-center">
-            <span className="font-semibold text-sm md:text-base">
-              Created On
-            </span>
-            <span
-              className=" px-3 py-1 rounded-xl text-xs md:text-sm w-32 text-center
-              bg-[rgba(232,232,232,1)] dark:bg-gray-700 dark:text-gray-100 "
-            >
-              {dayjs(project.Created_On).format("MMM D, YYYY")}
-            </span>
-          </div>
-
-          <div className="flex justify-between items-center">
-            <span className="font-semibold text-sm md:text-base">
-              Last Modified
-            </span>
-            <span
-              className="px-3 py-1 rounded-xl text-xs md:text-sm w-32 text-center
-              bg-[rgba(232,232,232,1)] dark:bg-gray-700 dark:text-gray-100"
-            >
-              {dayjs(project.Last_Modified).fromNow()}
-            </span>
-          </div>
-
-          <div className="flex justify-between items-center">
-            <span className="font-semibold text-sm md:text-base">Status</span>
-            <span className="bg-blue-600 text-white px-3 py-1 rounded-xl text-xs md:text-sm w-32 text-center">
-              {project.Status}
-            </span>
-          </div>
-
-          <div className="flex justify-between items-center">
-            <span className="font-semibold text-sm md:text-base">
-              Number of Points
-            </span>
-            <span
-              className="px-3 py-1 rounded-xl text-xs md:text-sm w-32 text-center
-              bg-[rgba(232,232,232,1)] dark:bg-gray-700 dark:text-gray-100"
-            >
-              {project.Total_Points || 0}
-            </span>
-          </div>
-
-
-          <div className="flex justify-between items-center">
-            <span className="font-semibold text-sm md:text-base">
-              Survey Time
-            </span>
-            <span
-              className="px-3 py-1 rounded-xl text-xs md:text-sm w-32 text-center
-              bg-[rgba(232,232,232,1)] dark:bg-gray-700 dark:text-gray-100"
-            >
-              {project.Survey_Time || "N/A"}
-            </span>
-          </div>
-        </div>
-
+        <ProjectOverview/>
 
         {/* /* assigned survayers */} 
 
@@ -366,73 +280,9 @@ const handleExport = async () => {
 
             {/* Sections Manage */}
           </div>
-         <div className="bg-white p-5 rounded-lg flex flex-col gap-2 h-max">
-              <h2 className="text-base md:text-lg font-semibold pb-2">Sections</h2>
-
-              <div className="flex flex-wrap gap-2">
-          {project.Sections && project.Sections.length > 0 ? (
-            project.Sections.map((section, idx) => (
-              <span
-                key={idx}
-                className="flex items-center bg-gray-200 dark:bg-gray-700 text-gray-800 dark:text-gray-100 px-3 py-1 rounded-full text-xs md:text-sm"
-              >
-                {section}
-                {section !== "default" && (
-  <button
-    className="ml-2 text-gray-900 hover:text-red-500 focus:outline-none text-lg"
-    title="Remove section"
-    onClick={() => {
-      const updatedSections = project.Sections.filter((_, i) => i !== idx);
-      setProject({
-        ...project,
-        Sections: updatedSections,
-      });
-      const removedSection = project.Sections[idx];
-      removeSection(projectId, removedSection);
-    }}
-  >
-    &times;
-  </button>
-)}
-
-              </span>
-            ))
-          ) : (
-            <span className="text-gray-500">No sections available</span>
-          )}
-              </div>
-
-              {/* Input + Button to add new section */}
-      <div className="mt-4 flex gap-2">
-        <input
-          type="text"
-          placeholder="Add new section"
-          className="flex-1 p-2 border rounded-lg text-sm md:text-base"
-          style={{ backgroundColor: "rgba(232, 232, 232, 1)" }}
-          value={newSection}
-          onChange={(e) => setNewSection(e.target.value)}
-          onKeyDown={(e) => {
-            if (e.key === "Enter") handleAddSection();
-          }}
-        />
-        <button
-          className="bg-black hover:bg-gray-900 text-white px-8 py-1.5 rounded-xl text-sm md:text-base
-              disabled:bg-gray-400 dark:bg-indigo-600 dark:hover:bg-indigo-500"
-          onClick={handleAddSection}
-        >
-          Add
-        </button>
-      </div>
-    </div>
-
+         <SectionsOverview/>
 
       </div>
-       
-           
-
-       
-        
-        
       </div>
       
     </div>
